+++ sqlagg/columns.py
@@ -1,13 +1,14 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
-from collections import OrderedDict
 from sqlalchemy import func, distinct, case, text, cast, Integer, column
+from sqlalchemy.dialects.postgresql import aggregate_order_by
 from .base import BaseColumn, CustomQueryColumn, SqlColumn
 import six
+import uuid
 
 
 class SimpleColumn(BaseColumn):
-    pass 
+    pass
 
 
 class YearColumn(BaseColumn):
@@ -74,9 +75,7 @@ class CountUniqueColumn(BaseColumn):
 class ConditionalAggregation(BaseColumn):
     def __init__(self, key=None, whens=None, else_=None, *args, **kwargs):
         super(ConditionalAggregation, self).__init__(key, *args, **kwargs)
-        # This appears in both the SELECT block and the GROUP BY block
-        # Until that changes, it must appear in a deterministic order
-        self.whens = OrderedDict(sorted((whens or {}).items()))
+        self.whens = whens or []
         self.else_ = else_
 
         assert self.key or self.alias, "Column must have either a key or an alias"
@@ -86,20 +85,57 @@ class ConditionalAggregation(BaseColumn):
         return ConditionalColumn(self.key, self.whens, self.else_, self.aggregate_fn, self.alias)
 
 
+class ArrayAggColumn(BaseColumn):
+    """
+    Perform array aggregation on a column
+    Pass order_by_col to sort by another column within the group
+    Example: array generated for a column col1, ordered by col2 with select clause like
+    Select ARRAY_AGG(col1 ORDER BY col2), col3 ..
+    """
+
+    def __init__(self, key, order_by_col=None, *args, **kwargs):
+        super(ArrayAggColumn, self).__init__(key, *args, **kwargs)
+        self.order_by_col = order_by_col
+
+    @property
+    def sql_column(self):
+        return ArrayAggSQLColumn(self.key, self.order_by_col, self.alias)
+
+
 class SumWhen(ConditionalAggregation):
     """
-    SumWhen("vehicle", whens={"unicycle": 1, "bicycle": 2, "car": 4}, else_=0, alias="num_wheels")
+    Without binds:
+    SumWhen("vehicle", whens=[["unicycle", 1], ["bicycle", 2], ["car", 4]], else_=0, alias="num_wheels")
+
+    With binds:
+    SumWhen("age_cohort",
+            whens=[
+                ["age_in_months < ?", 6, 1],
+                ["age_in_months < ?", 12, 2],
+            ],
+            else_=0,
+            alias="half_years")
     """
     aggregate_fn = func.sum
 
 
 class ConditionalColumn(SqlColumn):
     """
+    Without binds:
     ConditionalColumn("vehicle",
-                      whens={"unicycle": 1, "bicycle": 2, "car": 4},
+                      whens=[["unicycle", 1], ["bicycle": 2], ["car": 4]],
                       else_=0,
                       aggregation_fn=func.sum,
                       alias="num_wheels")
+
+    Or with binds:
+    ConditionalColumn("age_cohort",
+                      whens=[
+                        ["age_in_months < ?", 6, 1],
+                        ["age_in_months < ?", 12, 2],
+                      ],
+                      else_=0,
+                      aggregation_fn=func.sum)
     """
     def __init__(self, column_name, whens, else_, aggregate_fn, alias):
         self.aggregate_fn = aggregate_fn
@@ -116,15 +152,48 @@ class ConditionalColumn(SqlColumn):
         if self.column_name:
             expr = case(value=column(self.column_name), whens=self.whens, else_=self.else_)
         else:
-            whens = []
-            for when, then in self.whens.items():
-                if isinstance(then, six.string_types):
-                    whens.append((text(when), text(then)))
-                else:
-                    whens.append((text(when), then))
-
-            expr = case(whens=whens, else_=self.else_)
+            expr = case(whens=self._build_whens(), else_=self.else_)
 
         if self.aggregate_fn:
             expr = self.aggregate_fn(expr)
         return expr.label(self.label)
+
+    def _build_whens(self):
+        whens = []
+        for item in self.whens:
+            when, *binds, then = item
+            if binds:
+                binds = list(reversed(binds))
+                named_binds = {}
+                when_with_named_binds = ''
+                for letter in when:
+                    if letter != '?':
+                        when_with_named_binds += letter
+                    else:
+                        bind_name = 'p' + uuid.uuid4().hex
+                        when_with_named_binds += ':' + bind_name
+                        named_binds[bind_name] = binds.pop()
+                when = text(when_with_named_binds).bindparams(**named_binds)
+            else:
+                when = text(when)
+            then = text(then) if isinstance(then, six.string_types) else then
+            whens.append((when, then))
+        return whens
+
+
+class ArrayAggSQLColumn(SqlColumn):
+    def __init__(self, column_name, order_by_col, alias=None):
+        self.column_name = column_name
+        self.order_by_col = order_by_col
+        self.alias = alias
+
+    @property
+    def label(self):
+        return self.alias or self.column_name
+
+    def build_column(self):
+        table_column = column(self.column_name)
+        if self.order_by_col:
+            order_by_column = column(self.order_by_col)
+            return func.array_agg(aggregate_order_by(table_column, order_by_column.asc())).label(self.label)
+        return func.array_agg(table_column).label(self.label)