+++ CHANGES.md
@@ -1,6 +1,22 @@
 History
 =========
 
+0.16.0 (2020-03-24)
+----------------
+* Added support for distinct on (#66)
+
+0.15.0 (2020-01-21)
+----------------
+* Added support for array aggregation (#63)
+
+0.14.0 (2019-11-13)
+----------------
+* Added support for binds to SumWhen and ConditionalColumn (#60). Both of these classes now expect `whens` to be a list of lists rather than a dict.
+
+0.13.0 (2019-04-25)
+----------------
+* Removed SQL injection in OrderBy class (#57)
+
 0.1.1 (2013-05-24)
 ----------------
 * Added conditional columns (SumWhen)